Local branch amd-gfx 183042a5b335 Merged main:4c921aa3f554 into amd-gfx:7516e23b7471
Remote branch main acc9721e3866 [NFC] [HWASan] Remove indirection for getting analyses.

+++ polly/lib/Analysis/ScopInfo.cpp
@@ -2559,7 +2559,7 @@ void updateLoopCountStatistic(ScopDetection::LoopStats Stats,
   NumScops++;
   NumLoopsInScop += Stats.NumLoops;
   MaxNumLoopsInScop =
-      std::max(MaxNumLoopsInScop.getValue(), (unsigned)Stats.NumLoops);
+      std::max(MaxNumLoopsInScop.getValue(), (uint64_t)Stats.NumLoops);
 
   if (Stats.MaxDepth == 0)
     NumScopsDepthZero++;
